Move repository class to its own module
